[gaim-migrate @ 13521] Bits and pieces all over the place. I was hunting down compiler warnings all over the place, with crazy CFLAGS. I think I might have got a bit carried away.
I made the perl stuff compile with gcc 2.x (which don't support
mixed declerations and statements), I think what I did was safe but
readily admit I have no clue how all the perl stuff works.

Rename gaim_presence_add_presence() to gaim_presence_add_list() (to match
the header file - obviously nothing actually _uses_ this)

Implement gaim_savedstatus_get_handle() because it's in the header file.

Remove gaim_account_set_presence() from account.h - it's not actually
implemented anywhere, so it can't be that important.

--HG--
extra : convert_revision : 4c4f54c2503099a4e89acd5ff94bd4d61095c5ca

+++ src/protocols/irc/dcc_send.c
@@ -185,7 +185,7 @@ static void irc_dccsend_send_read(gpointer data, int source, GaimInputCondition 
 	xd->rxlen += len;
 
 	while (1) {
-		int acked;
+		size_t acked;
 
 		if (xd->rxlen < 4)
 			break;
